Add explicit `undefined` for optional properties
Typescript 4.4 (currently in beta) provides support for distinguishing
whether or not optional parameters should be allowed to be sete to an
undefined value.

In order to indicate optional values that support being set to
undefined, ` | undefined` must explicitly be included in the type
definition.

Also fix some doc strings regarding default values for map/visual styles